[FE] rename: redux hook 적용
- Slice 이름 변경 : createGroupDataSlice ->groupCreateDataSlice\
- Slice State 선언 및 export : groupCreateDataState
- useAppSelector, useAppDispatch 적용
