+++ client/src/pages/GroupCreatePage/TechStackInput/TechStackList.tsx
@@ -1,11 +1,9 @@
 import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { ReducerType } from '@store/rootReducer';
-import { setGroupData } from '@store/slices/createGroupData';
-import { GroupData } from '../../../types/CreateGroup';
+import { groupCreateDataState, setGroupData } from '@store/slices/groupCreateDataSlice';
 import { TechStack } from '@types';
+import { useAppDispatch, useAppSelector } from '@hooks';
 
 const MAX_SELECTED_INDEX = 5;
 
@@ -13,10 +11,8 @@ interface Props {
   techStackList: TechStack[];
 }
 function TechStackList({ techStackList }: Props): JSX.Element {
-  const { selectedTechStack } = useSelector<ReducerType, GroupData>(
-    (state) => state.createGroupData,
-  );
-  const dispatch = useDispatch();
+  const { selectedTechStack } = useAppSelector(groupCreateDataState);
+  const dispatch = useAppDispatch();
   const handleTechStackClick = (e: React.MouseEvent<HTMLButtonElement>) => {
     const clickedTechStack = e.target as HTMLButtonElement;
     const techStackName = clickedTechStack.innerText;
